Implement compilation of prefix operators
