+++ compiler/compiler.go
@@ -1,11 +1,7 @@
 package compiler
 
 import (
-	"fmt"
-	"reflect"
-
 	"github.com/Zac-Garby/pluto/ast"
-	"github.com/Zac-Garby/pluto/bytecode"
 	"github.com/Zac-Garby/pluto/object"
 )
 
@@ -35,120 +31,6 @@ func (c *Compiler) CompileProgram(p ast.Program) error {
 			return err
 		}
 	}
-
-	return nil
-}
-
-// CompileStatement compiles an AST statement.
-func (c *Compiler) CompileStatement(n ast.Statement) error {
-	switch node := n.(type) {
-	case *ast.ExpressionStatement:
-		return c.CompileExpression(node.Expr)
-	default:
-		return fmt.Errorf("compiler: compilation not yet implemented for %s", reflect.TypeOf(n))
-	}
-}
-
-// CompileExpression compiles an AST expression.
-func (c *Compiler) CompileExpression(n ast.Expression) error {
-	switch node := n.(type) {
-	case *ast.InfixExpression:
-		return c.compileInfix(node)
-	case *ast.Number:
-		return c.compileNumber(node)
-	case *ast.Identifier:
-		return c.compileIdentifier(node)
-	case *ast.AssignExpression:
-		return c.compileAssign(node)
-	default:
-		return fmt.Errorf("compiler: compilation not yet implemented for %s", reflect.TypeOf(n))
-	}
-}
-
-func (c *Compiler) compileNumber(node *ast.Number) error {
-	obj := &object.Number{Value: node.Value}
-	c.Constants = append(c.Constants, obj)
-	index := len(c.Constants) - 1
-
-	if index >= 1<<16 {
-		return fmt.Errorf("compiler: constant index %d greater than 1 << 16 (maximum uint16)", index)
-	}
-
-	low, high := runeToBytes(rune(index))
-
-	c.Bytes = append(c.Bytes, bytecode.LoadConst, high, low)
-
-	return nil
-}
-
-func (c *Compiler) compileIdentifier(node *ast.Identifier) error {
-	var index int
-
-	for i, name := range c.Names {
-		if name == node.Value {
-			index = i
-			goto found
-		}
-	}
-
-	// These two lines are executed if the name isn't found
-	c.Names = append(c.Names, node.Value)
-	index = len(c.Names) - 1
-
-found:
-	low, high := runeToBytes(rune(index))
-
-	c.Bytes = append(c.Bytes, bytecode.LoadName, high, low)
-
-	return nil
-}
-
-func (c *Compiler) compileAssign(node *ast.AssignExpression) error {
-	if err := c.CompileExpression(node.Value); err != nil {
-		return err
-	}
-
-	c.Names = append(c.Names, node.Name.(*ast.Identifier).Value)
-	index := len(c.Names) - 1
-
-	if index >= 1<<16 {
-		return fmt.Errorf("compiler: name index %d greater than 1 << 16 (maximum uint16)", index)
-	}
-
-	low, high := runeToBytes(rune(index))
-
-	c.Bytes = append(c.Bytes, bytecode.StoreName, high, low)
-
-	return nil
-}
-
-func (c *Compiler) compileInfix(node *ast.InfixExpression) error {
-	left, right := node.Left, node.Right
-
-	if err := c.CompileExpression(left); err != nil {
-		return err
-	}
-
-	if err := c.CompileExpression(right); err != nil {
-		return err
-	}
-
-	op := map[string]byte{
-		"+":  bytecode.BinaryAdd,
-		"-":  bytecode.BinarySubtract,
-		"*":  bytecode.BinaryMultiply,
-		"/":  bytecode.BinaryDivide,
-		"**": bytecode.BinaryExponent,
-		"//": bytecode.BinaryFloorDiv,
-		"%":  bytecode.BinaryFloorDiv,
-		"||": bytecode.BinaryOr,
-		"&&": bytecode.BinaryAnd,
-		"|":  bytecode.BinaryBitOr,
-		"&":  bytecode.BinaryBitAnd,
-		"==": bytecode.BinaryEquals,
-	}[node.Operator]
-
-	c.Bytes = append(c.Bytes, op)
 
 	return nil
 }
