+++ compiler/statements.go
@@ -1,0 +1,18 @@
+package compiler
+
+import (
+	"fmt"
+	"reflect"
+
+	"github.com/Zac-Garby/pluto/ast"
+)
+
+// CompileStatement compiles an AST statement.
+func (c *Compiler) CompileStatement(n ast.Statement) error {
+	switch node := n.(type) {
+	case *ast.ExpressionStatement:
+		return c.CompileExpression(node.Expr)
+	default:
+		return fmt.Errorf("compiler: compilation not yet implemented for %s", reflect.TypeOf(n))
+	}
+}
